requirements.txt
django>=4.2,<5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt>=5.2.0
drf-spectacular==0.26.5